boards: pico 2: split into library and binary
boards/raspberry_pi_pico is both a library and a binary: src/lib.rs holds
the platform every derived board shares, and src/main.rs is the thin
binary on top of it. boards/raspberry_pi_pico_w reuses it that way.
boards/raspberry_pi_pico_2 had only src/main.rs, so nothing could be
built on it.

Give it the same shape. src/lib.rs gains ChipHw, PANIC_RESOURCES,
SchedulerInUse, a Platform struct carrying the console, alarm, GPIO and
IPC drivers, and a setup() that brings the chip up and returns the
kernel, the platform, the peripherals, the alarm mux and the chip.
src/main.rs keeps the LED, the fault policy, the stack, process loading
and the main loop. No Cargo.toml change: cargo discovers src/lib.rs and
src/main.rs, which is how the pico family already works.

setup() returns the alarm mux because a derived board needs to hang
further virtual alarms off it; boards/raspberry_pi_pico does the same for
the same reason. Unlike that board there is no Output parameter, since
chips/rp2350 has no USB and the console is always UART0.

setup() does take one parameter, and it is the GPIO driver. Which pins
userspace may drive is a decision for each board rather than for the
platform, because a pin that is free on one may be wired to something
on another, so a derived board has to be able to choose. It cannot be
handed a finished driver: gpio_component_helper! fixes the pin set at
its call site, and building one needs board_kernel and peripherals,
which setup() is what creates. So setup() takes a closure that builds
the driver from those two, and the pin list moves out of lib.rs into
each binary. raspberry_pi_pico_2's own list moves unchanged, commented
out pins and all, so this board exposes exactly the pins it did before.

The panic handler stays in each binary's io.rs, so the WRITER.set_uart
call moves out of setup() and runs once it returns. This is not a
behaviour change. WRITER is read only by panic_fmt, and when it is unset
io.rs builds a Uart over the same registers and calls configure_uart,
which gates on Uart::is_configured(). That reads UARTEN and RXE or TXE
out of UARTCR, so it asks the hardware rather than a software flag, and
finds the UART setup() already configured. Both paths leave the same
peripheral in the same state.

    text    data     bss
    69164      0   19100   before
    69164      0   19104   after

Flash does not move. The four bytes of bss do, and I have not pinned
down which allocation they are; the figure is measured, the cause is
not, so it is left as an observation rather than an explanation.

The stripped binary is not byte identical, and cannot be. Moving the code
into a library crate changes the partition LLVM starts from, so inlining
and machine outlining land differently: 84 outlined functions become 86,
and individual functions move a few bytes either way. The section totals
absorb it exactly. METADATA_BLOCK and FLASH_BOOTLOADER also change from
local to global symbols, which is what compiling them into a library
means; their addresses and sizes do not move.

The boot image was checked rather than assumed, since dropping either
blob links cleanly and only fails on hardware. Both are present at the
right addresses and unchanged: the metadata block's ffffded3 magic at
0x10000000 with its VECTOR_TABLE item still 0x10000400, and boot2 at
0x1000001c, 256 bytes.

Nothing here was run on hardware. That this still boots is a regression
check for the bench session that comes with the Pico 2 W board.
